pretrain/requirements.txt
numpy==1.22.0
torchvision==0.8.1
psutil==5.6.7
tqdm==4.48.2
torch==1.7.0
pillow>=8.3.2
scikit_learn==0.24.1
torchscan==0.1.1
efficientnet_pytorch==0.7.1
pandas>=1.0.0
